+++ Math.h
@@ -10,7 +10,7 @@
 #ifndef B2BSAND_MATH_H
 #define B2BSAND_MATH_H
 
-int randomBetween(int32_t _low, uint64_t _high);
+int32_t randomBetween(int32_t _low, int32_t _high);
 bool intersectRect(const Position* p1, const Dimension* d1, const Position* p2, const Dimension* d2);
 
 #endif //B2BSAND_MATH_H
